python-service/requirements.txt
# FastAPI-based Resume Analyzer Service Dependencies
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# File processing
PyPDF2==3.0.1
python-docx==1.1.0

# Database
motor==3.3.2
pymongo==4.6.0

# AI/ML
groq==0.4.1

# Validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0 